BoolUnification3 -> BoolUnification
